INDY-1454: Refactor tests to use safe context managers in tconf
