+++ plenum/test/view_change_with_delays/test_two_view_changes_with_delayed_commits.py
@@ -1,6 +1,6 @@
 import pytest
 
-from plenum.test.helper import sdk_send_random_and_check
+from plenum.test.helper import sdk_send_random_and_check, perf_monitor_disabled
 from plenum.test.view_change_with_delays.helper import do_view_change_with_pending_request_and_one_fast_node
 
 # This is needed only with current view change implementation to give enough time
@@ -13,10 +13,8 @@ def tconf(tconf):
     """
     Patch config so that monitor won't start view change unexpectedly
     """
-    old_unsafe = tconf.unsafe
-    tconf.unsafe.add("disable_view_change")
-    yield tconf
-    tconf.unsafe = old_unsafe
+    with perf_monitor_disabled(tconf):
+        yield tconf
 
 
 @pytest.mark.skip(reason="INDY-1303, case 2")
